+++ evil-icons.js
@@ -1,31 +1,7 @@
 var fs          = require('fs');
 var spritePath  = __dirname + '/assets/sprite.svg';
 var sprite      = fs.readFileSync(spritePath).toString();
-
-function icon(name, options) {
-  var options = options || {};
-  var size    = options.size ? ' icon--' + options.size : '';
-  var classes = 'icon icon--' + name + size + ' ' + (options.class || '');
-  classes     = classes.trim();
-
-  var icon =  '<svg class="icon__cnt">' +
-                '<use xlink:href="#' + name + '-icon" />' +
-              '</svg>';
-
-  var html =  '<div class="' + classes + '">' +
-                wrapSpinner(icon, classes) +
-              '</div>';
-
-  return html;
-}
-
-function wrapSpinner(html, klass) {
-  if (klass.indexOf('spinner') > -1) {
-    return '<div class="icon__spinner">' + html + '</div>';
-  } else {
-    return html;
-  }
-}
+var icon        = require('./icon');;
 
 function buildParamsFromString(string) {
   var match, attr, value;
